+++ Podfile
@@ -1,0 +1,20 @@
+# Uncomment the next line to define a global platform for your project
+ platform :ios, '10.0'
+
+target 'mcheckers' do
+  # Comment the next line if you're not using Swift and don't want to use dynamic frameworks
+  use_frameworks!
+
+  # Pods for mcheckers
+
+  target 'mcheckersTests' do
+    inherit! :search_paths
+    # Pods for testing
+  end
+
+  target 'mcheckersUITests' do
+    inherit! :search_paths
+    # Pods for testing
+  end
+
+end
