+++ mcheckers/Other/Piece.swift
@@ -1,0 +1,16 @@
+//
+//  Piece.swift
+//  mcheckers
+//
+//  Created by Melissa Golson on 6/3/17.
+//  Copyright © 2017 bgsoft. All rights reserved.
+//
+
+import Foundation
+
+class Piece {
+    var owner = "p1"
+    var location = "A2"
+    var kinged = false
+    var score = 2
+}
