+++ mcheckers/Other/Square.swift
@@ -1,0 +1,16 @@
+//
+//  Square.swift
+//  mcheckers
+//
+//  Created by Melissa Golson on 6/2/17.
+//  Copyright © 2017 bgsoft. All rights reserved.
+//
+
+import Foundation
+
+class Square {
+    var color = "white"
+    var location = "A1"
+    var occupied = true
+}
+
