--- mcheckers/PieceClass.swift
@@ -1,103 +1,0 @@
-//
-//  PieceClass.swift
-//  mcheckers
-//
-//  Created by Melissa Golson on 6/3/17.
-//  Copyright © 2017 bgsoft. All rights reserved.
-//
-
-import Foundation
-
-class Piece {
-    var owner = "p1"
-    var location = "A2"
-    var kinged = false
-    var score = 2
-}
-
-//Player 1 pieces
-
-let p1Piece1 = Piece()
-
-let p1Piece2 = Piece()
-p1Piece2.location = "A4"
-
-let p1Piece3 = Piece()
-p1Piece3.location = "A6"
-
-let p1Piece4 = Piece()
-p1Piece4.location = "A8"
-
-let p1Piece5 = Piece()
-p1Piece5.location = "B1"
-
-let p1Piece6 = Piece()
-p1Piece6.location = "B3"
-
-let p1Piece7 = Piece()
-p1Piece7.location = "B5"
-
-let p1Piece8 = Piece()
-p1Piece8.location = "B7"
-
-let p1Piece9 = Piece()
-p1Piece9.location = "C2"
-
-let p1Piece10 = Piece()
-p1Piece10.location = "C4"
-
-let p1Piece11 = Piece()
-p1Piece11.location = "C6"
-
-let p1Piece12 = Piece()
-p1Piece12.location = "C8"
-
-//Player 2 pieces
-
-let p2Piece1 = Piece()
-p2Piece1.owner = "p2"
-p2Piece1.location = "F1"
-
-let p2Piece2 = Piece()
-p2Piece2.owner = "p2"
-p2Piece2.location = "F3"
-
-let p2Piece3 = Piece()
-p2Piece3.owner = "p2"
-p2Piece3.location = "F5"
-
-let p2Piece4 = Piece()
-p2Piece4.owner = "p2"
-p2Piece4.location = "F7"
-
-let p2Piece5 = Piece()
-p2Piece5.owner = "p2"
-p2Piece5.location = "G2"
-
-let p2Piece6 = Piece()
-p2Piece6.owner = "p2"
-p2Piece6.location = "G4"
-
-let p2Piece7 = Piece()
-p2Piece7.owner = "p2"
-p2Piece7.location = "G6"
-
-let p2Piece8 = Piece()
-p2Piece8.owner = "p2"
-p2Piece8.location = "G8"
-
-let p2Piece9 = Piece()
-p2Piece9.owner = "p2"
-p2Piece9.location = "H1"
-
-let p2Piece10 = Piece()
-p2Piece10.owner = "p2"
-p2Piece10.location = "H3"
-
-let p2Piece11 = Piece()
-p2Piece11.owner = "p2"
-p2Piece11.location = "H5"
-
-let p2Piece12 = Piece()
-p2Piece12.owner = "p2"
-p2Piece12.location = "H7"
