--- mcheckers/SquareClass.swift
@@ -1,278 +1,0 @@
-//
-//  SquareClass.swift
-//  mcheckers
-//
-//  Created by Melissa Golson on 6/2/17.
-//  Copyright © 2017 bgsoft. All rights reserved.
-//
-
-import Foundation
-
-class Square {
-    var color = "white"
-    var location = "A1"
-    var occupied = true
-}
-
-let square1 = Square()
-square1.occupied = false
-
-let square2 = Square()
-square2.color = "black"
-square2.location = "A2"
-
-let square3 = Square()
-square3.location = "A3"
-square3.occupied = false
-
-let square4 = Square()
-square4.color = "black"
-square4.location = "A4"
-
-let square5 = Square()
-square5.location = "A5"
-square5.occupied = false
-
-let square6 = Square()
-square6.color = "black"
-square6.location = "A6"
-
-let square7 = Square()
-square7.location = "A7"
-square7.occupied = false
-
-let square8 = Square()
-square8.color = "black"
-square8.location = "A8"		//end of first row
-
-let square9 = Square()
-square9.color = "black"
-square9.location = "B1"
-
-let square10 = Square()
-square10.location = "B2"
-square10.occupied = false
-
-let square11 = Square()
-square11.color = "black"
-square11.location = "B3"
-
-let square12 = Square()
-square12.location = "B4"
-square12.occupied = false
-
-let square13 = Square()
-square13.color = "black"
-square13.location = "B5"
-
-let square14 = Square()
-square14.location = "B6"
-square14.occupied = false
-
-let square15 = Square()
-square15.color = "black"
-square15.location = "B7"
-
-let square16 = Square()
-square16.location = "B8"
-square16.occupied = false	//end of second row
-
-let square17 = Square()
-square17.location = "C1"
-square17.occupied = false
-
-let square18 = Square()
-square18.color = "black"
-square18.location = "C2"
-
-let square19 = Square()
-square19.location = "C3"
-square19.occupied = false
-
-let square20 = Square()
-square20.color = "black"
-square20.location = "C4"
-
-let square21 = Square()
-square21.location = "C5"
-square21.occupied = false
-
-let square22 = Square()
-square22.color = "black"
-square22.location = "C6"
-
-let square23 = Square()
-square23.location = "C7"
-square23.occupied = false
-
-let square24 = Square()
-square24.color = "black"
-square24.location = "C8"		//end of third row
-
-let square25 = Square()
-square25.color = "black"
-square25.location = "D1"
-square25.occupied = false
-
-let square26 = Square()
-square26.location = "D2"
-square26.occupied = false
-
-let square27 = Square()
-square27.color = "black"
-square27.location = "D3"
-square27.occupied = false
-
-let square28 = Square()
-square28.location = "D4"
-square28.occupied = false
-
-let square29 = Square()
-square29.color = "black"
-square29.location = "D5"
-square.occupied = false
-
-let square30 = Square()
-square30.location = "D6"
-square30.occupied = false
-
-let square31 = Square()
-square31.color = "black"
-square31.location = "D7"
-square31.occupied = false
-
-let square32 = Square()
-square32.location = "D8"
-square32.occupied = false	//end of fourth row
-
-let square33 = Square()
-square33.location = "E1"
-square33.occupied = false
-
-let square34 = Square()
-square34.color = "black"
-square34.location = "E2"
-square34.occupied = false
-
-let square35 = Square()
-square35.location = "E3"
-square36.occupied = false
-
-let square36 = Square()
-square36.color = "black"
-square36.location = "E4"
-square36.occupied = false
-
-let square37 = Square()
-square37.location = "E5"
-square37.occupied = false
-
-let square38 = Square()
-square38.color = "black"
-square38.location = "E6"
-square38.occupied = false
-
-let square39 = Square()
-square39.location = "E7"
-square39.occupied = false
-
-let square40 = Square()
-square40.color = "black"
-square40.location = "E8"
-square40.occupied = false	//end of fifth row
-
-let square41 = Square()
-square41.color = "black"
-square41.location = "F1"
-
-let square42 = Square()
-square42.location = "F2"
-square42.occupied = false
-
-let square43 = Square()
-square43.color = "black"
-square43.location = "F3"
-
-let square44 = Square()
-square44.location = "F4"
-square44.occupied = false
-
-let square45 = Square()
-square45.color = "black"
-square45.location = "F5"
-
-let square46 = Square()
-square46.location = "F6"
-square46.occupied = false
-
-let square47 = Square()
-square47.color = "black"
-square47.location = "F7"
-
-let square48 = Square()
-square48.location = "F8"
-square48.occupied = false		//end of sixth row
-
-let square49 = Square()
-square49.location = "G1"
-square49.occupied = false
-
-let square50 = Square()
-square50.color = "black"
-square50.location = "G2"
-
-let square51 = Square()
-square51.location = "G3"
-square51.occupied = false
-
-let square52 = Square()
-square52.color = "black"
-square52.location = "G4"
-
-let square53 = Square()
-square53.location = "G5"
-square53.occupied = false
-
-let square54 = Square()
-square54.color = "black"
-square54.location = "G6"
-
-let square55 = Square()
-square55.location = "G7"
-square55.occupied = false
-
-let square56 = Square()
-square56.color = "black"
-square56.location = "G8"		//end of seventh row
-
-let square57 = Square()
-square57.color = "black"
-square57.location = "H1"
-
-let square58 = Square()
-square58.location = "H2"
-square58.occupied = false
-
-let square59 = Square()
-square59.color = "black"
-square59.location = "H3"
-
-let square60 = Square()
-square60.location = "H4"
-square60.occupied = false
-
-let square61 = Square()
-square61.color = "black"
-square61.location = "H5"
-
-let square62 = Square()
-square62.location = "H6"
-square62.occupied = false
-
-let square63 = Square()
-square63.color = "black"
-square63.location = "H7"
-
-let square64 = Square()
-square64.location = "H8"
-square64.occupied = false	//end of eighth row
